Conflicts:
	mysql-test/suite/i_innodb/t/disabled.def
	mysql-test/collections/default.push
	mysql-test/include/excludenoskip.list
	mysql-test/r/create_ps_protocol.result
	mysql-test/r/dd_schema_definition_debug_ci.result
	mysql-test/r/dd_schema_definition_debug_cs.result
	mysql-test/r/dd_stages.result
	mysql-test/r/group_min_max_ps_protocol.result
	mysql-test/r/mysqld--help-win.result
	mysql-test/r/partition_explicit_prune.result
	mysql-test/r/partition_locking.result
	mysql-test/r/partition_locking_ps_protocol.result
	mysql-test/suite/funcs_1/r/is_statistics.result
	mysql-test/suite/funcs_1/r/is_statistics_ci.result
	mysql-test/suite/innodb/r/create_tablespace_32k.result
	mysql-test/suite/innodb/r/create_tablespace_4k.result
	mysql-test/suite/innodb/r/create_tablespace_64k.result
	mysql-test/suite/innodb/r/create_tablespace_8k.result
	mysql-test/suite/innodb/r/innodb-system-table-view_cs.result
	mysql-test/suite/innodb/r/innodb.result
	mysql-test/suite/innodb_zip/r/16k.result
	mysql-test/suite/perfschema/r/alter_table_progress.result
	mysql-test/suite/perfschema/t/disabled.def
	mysql-test/suite/perfschema/t/statement_digest_long_query.test
	mysql-test/t/disabled.def
	sql/binlog.cc
	sql/dd/dd_table.cc
	sql/dd/impl/system_registry.cc
	sql/item_create.cc
	sql/sql_partition_admin.cc
	sql/sql_table.cc
	sql/sql_truncate.cc
	storage/innobase/api/api0api.cc
	storage/innobase/dict/dict0dict.cc
	storage/innobase/dict/dict0load.cc
	storage/innobase/dict/dict0stats_bg.cc
	storage/innobase/fsp/fsp0sysspace.cc
	storage/innobase/fts/fts0fts.cc
	storage/innobase/fts/fts0opt.cc
	storage/innobase/handler/ha_innodb.cc
	storage/innobase/handler/ha_innopart.cc
	storage/innobase/handler/handler0alter.cc
	storage/innobase/handler/i_s.cc
	storage/innobase/include/dict0mem.h
	storage/innobase/log/log0recv.cc
	storage/innobase/pars/pars0pars.cc
	storage/innobase/row/row0ftsort.cc
	storage/innobase/row/row0import.cc
	storage/innobase/row/row0ins.cc
	storage/innobase/row/row0sel.cc
	storage/innobase/row/row0upd.cc
	storage/innobase/srv/srv0srv.cc
	storage/innobase/srv/srv0start.cc
	storage/innobase/sync/sync0debug.cc
	storage/innobase/trx/trx0trx.cc

+++ client/dump/sql_formatter.cc
@@ -70,7 +70,13 @@ void Sql_formatter::format_row_group(Row_group_dump_task* row_group)
 
   if (m_options->m_insert_type_replace)
     row_string+= "REPLACE INTO ";
-  else if (m_options->m_insert_type_ignore)
+  /*
+   for mysql.innodb_table_stats, mysql.innodb_index_stats tables always
+   dump as INSERT IGNORE INTO
+  */
+  else if (m_options->m_insert_type_ignore ||
+    innodb_stats_tables(row_group->m_source_table->get_schema(),
+                        row_group->m_source_table->get_name()))
     row_string+= "INSERT IGNORE INTO ";
   else
     row_string+= "INSERT INTO ";
@@ -213,6 +219,13 @@ void Sql_formatter::format_table_definition(
   Table_definition_dump_task* table_definition_dump_task)
 {
   Table* table= table_definition_dump_task->get_related_table();
+
+  /*
+   do not dump DDLs for mysql.innodb_table_stats,
+   mysql.innodb_index_stats tables
+  */
+  if (innodb_stats_tables(table->get_schema(), table->get_name()))
+    return;
   bool use_added= false;
   if (m_options->m_drop_table)
     this->append_output("DROP TABLE IF EXISTS "
@@ -264,7 +277,9 @@ void Sql_formatter::format_dump_end(Dump_end_dump_task* dump_start_dump_task)
   out << "SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;\n"
     "SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;\n"
     "SET SQL_MODE=@OLD_SQL_MODE;\n";
-
+  if (m_options->m_innodb_stats_tables_included)
+    out << "SET GLOBAL INNODB_STATS_AUTO_RECALC="
+      << "@OLD_INNODB_STATS_AUTO_RECALC;\n";
   out << "-- Dump end time: " << time_string << "\n";
 
   this->append_output(out.str());
@@ -291,6 +306,8 @@ void Sql_formatter::format_dump_start(
     "FOREIGN_KEY_CHECKS=0;\n" << "SET @OLD_SQL_MODE=@@SQL_MODE;\n"
     "SET SQL_MODE=\"NO_AUTO_VALUE_ON_ZERO\";\n";
 
+  /* disable binlog */
+  out << "SET @@SESSION.SQL_LOG_BIN= 0;\n";
 
   if (m_options->m_timezone_consistent)
     out << "SET @OLD_TIME_ZONE=@@TIME_ZONE;\n"
@@ -302,6 +319,44 @@ void Sql_formatter::format_dump_start(
     "SET NAMES "
     << this->get_charset()->csname
     << ";\n";
+
+  if (m_options->m_innodb_stats_tables_included)
+    out << "SET @OLD_INNODB_STATS_AUTO_RECALC="
+      << "@@INNODB_STATS_AUTO_RECALC;\n"
+      << "SET GLOBAL INNODB_STATS_AUTO_RECALC=OFF;\n";
+
+  if (dump_start_dump_task->m_gtid_mode == "OFF" &&
+      *((ulong*)&m_options->m_gtid_purged) == ((ulong)GTID_PURGED_ON))
+  {
+    m_options->m_mysql_chain_element_options->get_program()->error(
+      Mysql::Tools::Base::Message_data(1, "Server has GTIDs disabled.\n",
+      Mysql::Tools::Base::Message_type_error));
+    return;
+  }
+  if (dump_start_dump_task->m_gtid_mode != "OFF")
+  {
+    /*
+     value for m_gtid_purged is set by typecasting its address to ulong*
+     however below conditions fails if we do direct comparison without
+     typecasting on solaris sparc. Guessing that this is due to differnt
+     endianess.
+    */
+    if (*((ulong*)&m_options->m_gtid_purged) == ((ulong)GTID_PURGED_ON) ||
+        *((ulong*)&m_options->m_gtid_purged) == ((ulong)GTID_PURGED_AUTO))
+    {
+      if (!m_mysqldump_tool_options->m_dump_all_databases)
+      {
+        m_options->m_mysql_chain_element_options->get_program()->error(
+          Mysql::Tools::Base::Message_data(1,
+          "A partial dump from a server that has GTIDs is not allowed.\n",
+          Mysql::Tools::Base::Message_type_error));
+        return;
+      }
+      std::string gtid_output("SET @@GLOBAL.GTID_PURGED=/*!80000 '+'*/ '");
+      gtid_output+= (dump_start_dump_task->m_gtid_executed + "';\n");
+      out << gtid_output;
+    }
+  }
 
   this->append_output(out.str());
 }
@@ -384,6 +439,24 @@ void Sql_formatter::format_sql_objects_definer(
     plain_sql_dump_task->set_sql_formatted_definition(new_sql_stmt);
   }
 }
+
+/**
+  Check if the table is innodb stats table in mysql database.
+
+   @param [in] db           Database name
+   @param [in] table        Table name
+
+  @return
+    @retval TRUE if it is innodb stats table else FALSE
+*/
+bool Sql_formatter::innodb_stats_tables(std::string db,
+                                        std::string table)
+{
+  return ((db == "mysql") &&
+          ((table == "innodb_table_stats") ||
+          (table == "innodb_index_stats")));
+}
+
 void Sql_formatter::format_object(Item_processing_data* item_to_process)
 {
   this->object_processing_starts(item_to_process);
@@ -419,11 +492,13 @@ void Sql_formatter::format_object(Item_processing_data* item_to_process)
 Sql_formatter::Sql_formatter(I_connection_provider* connection_provider,
   std::function<bool(const Mysql::Tools::Base::Message_data&)>*
     message_handler, Simple_id_generator* object_id_generator,
+  const Mysqldump_tool_chain_maker_options* mysqldump_tool_options,
   const Sql_formatter_options* options)
   : Abstract_output_writer_wrapper(message_handler, object_id_generator),
   Abstract_mysql_chain_element_extension(
   connection_provider, message_handler,
   options->m_mysql_chain_element_options),
+  m_mysqldump_tool_options(mysqldump_tool_options),
   m_options(options)
 {
   m_escaping_runner= this->get_runner();
